backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
playwright==1.44.0
beautifulsoup4==4.12.3
requests==2.32.3
ollama==0.2.1
firebase-admin==6.5.0
python-dotenv==1.0.1
fake-useragent==1.5.1
lxml==5.2.2
python-multipart==0.0.9
httpx==0.27.0
slowapi==0.1.9